# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chronivaro REST API
  description: |
    REST API specification for Chronivaro, a Resource-Order-Activity time tracking and absence management system built on Strolch.
    Provides endpoints for time tracking, timers, presence overview, absence requests, period closing, reports, and administrative management.
  version: 0.1.0-SNAPSHOT
  contact:
    name: Chronivaro
    url: https://github.com/eitch/Chronivaro
servers:
  - url: /
    description: Current server instance
paths:
...
  /chronivaro/v1/absences/{id}/approve:
    post:
      summary: Approve absence
      description: Approves a pending absence request (supervisor/admin action). Supports optimistic concurrency via `If-Match`.
      operationId: approveAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      responses:
        '200':
          description: Absence approved successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/absences/{id}/reject:
    post:
      summary: Reject absence
      description: Rejects a pending absence request with an optional comment. Supports optimistic concurrency via `If-Match`.
      operationId: rejectAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionCommentDto'
      responses:
        '200':
          description: Absence rejected successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /chronivaro/v1/periods/approve:
    post:
      summary: Approve monthly period
      description: Approves and locks a submitted monthly period (supervisor/admin action).
      operationId: approvePeriod
      tags:
        - Periods
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeriodActionRequestDto'
      responses:
        '200':
          description: Period approved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodStatusDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /chronivaro/v1/admin/audit-logs:
    get:
      summary: Query system audit logs
      description: Returns immutable audit log records with filtering by entity, user, date range, or action.
      operationId: getAuditLogs
      tags:
        - Admin - Audits & Configuration
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: entityType
          in: query
          required: false
          description: Entity type filter (e.g., Employee, Absence, WorkEntry)
          schema:
            type: string
        - name: entityId
          in: query
          required: false
          description: Specific entity identifier
          schema:
            type: string
        - name: username
          in: query
          required: false
          description: Performed by username
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Earliest audit log timestamp (ISO)
          schema:
            type: string
        - name: to
          in: query
          required: false
          description: Latest audit log timestamp (ISO)
          schema:
            type: string
      responses:
        '200':
          description: Audit logs retrieved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/AuditLogDto'
                  - $ref: '#/components/schemas/PagedAuditLogDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: strolch.authorization
      description: Strolch session cookie containing the user authentication token.
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token containing the Strolch user certificate ID.
  headers:
    ETag:
      description: Entity version validator for optimistic concurrency control (e.g., W/"1" or "1").
      schema:
        type: string
        example: '"1"'
...
    X-Correlation-Id:
      description: Unique identifier for tracking and correlating requests across logs and client telemetry.
      schema:
        type: string
        format: uuid
        example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Zero-based item offset for pagination (defaults to 0).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
        example: 0
    LimitParam:
      name: limit
      in: query
      description: Number of items to return (defaults to 50, maximum 1000).
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
        example: 50
    IfMatchParam:
      name: If-Match
      in: header
      required: false
      description: Target entity version for optimistic concurrency control.
      schema:
        type: string
        example: '"1"'
  responses:
    BadRequest:
      description: Bad request - Validation failure or malformed payload
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Unauthorized:
      description: Unauthorized - Authentication is missing or invalid
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Forbidden:
      description: Forbidden - Insufficient user privileges or role permissions
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    NotFound:
      description: Not found - Requested resource does not exist
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Conflict:
      description: Conflict - Concurrency mismatch (stale If-Match) or business state violation
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    InternalServerError:
      description: Internal server error - Unexpected system failure
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
  schemas:
...
    ErrorDto:
      type: object
      required:
        - error
        - code
        - message
        - correlationId
      properties:
        error:
          type: boolean
          example: true
        code:
          type: string
          example: 'VALIDATION_ERROR'
        message:
          type: string
          example: 'Invalid input parameters'
        correlationId:
          type: string
          example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    FieldErrorDto:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: 'email'
        message:
          type: string
          example: 'Email address is invalid'
...
    ActionCommentDto:
      type: object
      properties:
        comment:
          type: string
          example: 'Approved based on team coverage agreement'
...
    AbsenceDto:
      type: object
      properties:
        id:
          type: string
          example: 'abs-001'
        employeeId:
          type: string
          example: 'emp-001'
        employeeName:
          type: string
          nullable: true
          example: 'John Doe'
        personalNumber:
          type: string
          nullable: true
          example: 'PER-1001'
        teamName:
          type: string
          nullable: true
          example: 'Engineering'
        absenceTypeCode:
          type: string
          example: 'VACATION'
        absenceTypeName:
          type: string
          nullable: true
          example: 'Vacation'
        start:
          type: string
          format: date-time
          example: '2026-08-20T00:00:00+02:00'
        end:
          type: string
          format: date-time
          example: '2026-08-22T23:59:59+02:00'
        durationType:
          type: string
          enum: [DAYS, HALF_DAY, HOURS]
          example: 'DAYS'
        dayPart:
          type: string
          enum: [MORNING, AFTERNOON]
          nullable: true
          example: 'MORNING'
        minutes:
          type: integer
          nullable: true
          example: 1440
        comment:
          type: string
          nullable: true
          example: 'Summer vacation'
        state:
          type: string
          enum: [REQUESTED, APPROVED, REJECTED, CANCELLED]
          example: 'APPROVED'
...
    PeriodStatusDto:
      type: object
      properties:
        employeeId:
          type: string
          example: 'emp-001'
        employeeName:
          type: string
          nullable: true
          example: 'John Doe'
        personalNumber:
          type: string
          nullable: true
          example: 'PER-1001'
        teamName:
          type: string
          nullable: true
          example: 'Engineering'
        yearMonth:
          type: string
          example: '2026-08'
        status:
          type: string
          enum: [OPEN, SUBMITTED, APPROVED, CLOSED, REJECTED, LOCKED]
          example: 'APPROVED'
        submittedAt:
          type: string
          format: date-time
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        approvedBy:
          type: string
          nullable: true
          example: 'supervisor1'
        rejectedAt:
          type: string
          format: date-time
          nullable: true
        rejectedBy:
          type: string
          nullable: true
          example: 'supervisor1'
        comment:
          type: string
          nullable: true
          example: 'Monthly review completed'
        calculationSnapshot:
          type: string
          nullable: true
    PeriodActionRequestDto:
      type: object
      required:
        - employeeId
        - yearMonth
      properties:
        employeeId:
          type: string
          example: 'emp-001'
        yearMonth:
          type: string
          example: '2026-08'
        comment:
          type: string
          example: 'Monthly review completed'
...
    AuditLogDto:
      type: object
      properties:
        id:
          type: string
          example: 'audit-001'
        timestamp:
          type: string
          format: date-time
          example: '2026-08-18T10:30:00+02:00'
        username:
          type: string
          example: 'admin'
        action:
          type: string
          enum: [CREATE, UPDATE, DELETE, SUBMIT, APPROVE, REJECT, CANCEL, CLOSE]
          example: 'CREATE'
        entityType:
          type: string
          example: 'Employee'
        entityId:
          type: string
          example: 'emp-001'
        details:
          type: string
          example: 'Created employee John Doe'
        reason:
          type: string
          example: 'Schedule updated'
        paramName:
          type: string
          example: 'weeklyTargetMinutes'
        oldValue:
          type: string
          example: '2400'
        newValue:
          type: string
          example: '2520'
        correlationId:
          type: string
          example: 'corr-12345'
    PagedAuditLogDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogDto'
        offset:
          type: integer
          example: 0
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 100
        size:
          type: integer
          example: 50
...